+++ Cura/util/profile.py
@@ -178,8 +178,6 @@ setting('solid_layer_thickness',     0.6, float, 'basic',    _('Fill')).setRange
 setting('fill_density',               20, float, 'basic',    _('Fill')).setRange(0, 100).setLabel(_("Fill Density (%)"), _("This controls how densely filled the insides of your print will be. For a solid part use 100%, for an empty part use 0%. A value around 20% is usually enough.\nThis won't affect the outside of the print and only adjusts how strong the part becomes."))
 setting('nozzle_size',               0.4, float, 'machine', _('Machine')).setRange(0.1,10).setLabel(_("Nozzle size (mm)"), _("The nozzle size is very important, this is used to calculate the line width of the infill, and used to calculate the amount of outside wall lines and thickness for the wall thickness you entered in the print settings."))
 setting('print_speed',                50, float, 'basic',    _('Speed and Temperature')).setRange(1).setLabel(_("Print speed (mm/s)"), _("Speed at which printing happens. A well adjusted Ultimaker can reach 150mm/s, but for good quality prints you want to print slower. Printing speed depends on a lot of factors. So you will be experimenting with optimal settings for this."))
-setting('grip_temperature',          220, int,   'basic',    _('Speed and Temperature')).setRange(0,340).setLabel(_("Grip temperature (C)"), _("Temperature used for printing. Set at 0 to pre-heat yourself.\nFor PLA a value of 210C is usually used.\nFor ABS a value of 230C or higher is required."))
-setting('grip_temperature2',          220, int,   'basic',    _('Speed and Temperature')).setRange(0,340).setLabel(_("Grip temperature (C)"), _("Temperature used for printing. Set at 0 to pre-heat yourself.\nFor PLA a value of 210C is usually used.\nFor ABS a value of 230C or higher is required."))
 setting('print_temperature',         220, int,   'basic',    _('Speed and Temperature')).setRange(0,340).setLabel(_("Printing temperature (C)"), _("Temperature used for printing. Set at 0 to pre-heat yourself.\nFor PLA a value of 210C is usually used.\nFor ABS a value of 230C or higher is required."))
 setting('print_temperature2',        220, int,   'basic',    _('Speed and Temperature')).setRange(0,340).setLabel(_("2nd nozzle temperature (C)"), _("Temperature used for printing. Set at 0 to pre-heat yourself.\nFor PLA a value of 210C is usually used.\nFor ABS a value of 230C or higher is required."))
 setting('print_temperature3',          0, int,   'basic',    _('Speed and Temperature')).setRange(0,340).setLabel(_("3th nozzle temperature (C)"), _("Temperature used for printing. Set at 0 to pre-heat yourself.\nFor PLA a value of 210C is usually used.\nFor ABS a value of 230C or higher is required."))
@@ -1248,7 +1246,6 @@ def getAlterationFileContents(filename, extruderCount = 1):
 
 def printSlicingInfo():
 	print '********* Slicing parameters *********'
-	print "grip_temperature : ", getProfileSetting('grip_temperature')
 	print "print_temperature : ", getProfileSetting('print_temperature')
 	print "filament_diameter : ", getProfileSetting('filament_diameter')
 	print "filament_flow : ", getProfileSetting('filament_flow')
@@ -1257,7 +1254,6 @@ def printSlicingInfo():
 	print "filament_physical_density : ", getProfileSetting('filament_physical_density')
 	print "filament_cost_kg : ", getProfileSetting('filament_cost_kg')
 	if int(getMachineSetting('extruder_amount')) == 2:
-		print "grip_temperature2 : ", getProfileSetting('grip_temperature2')
 		print "print_temperature2 : ", getProfileSetting('print_temperature2')
 		print "filament_diameter2 : ", getProfileSetting('filament_diameter2')
 		print "filament_flow2 : ", getProfileSetting('filament_flow2')